alttp: unify syncableItems[] on uint32 offset values in either wram 0..efff or sram f000..f500 ranges; unify 8-bit vs 16-bit asm generation

+++ games/alttp/frames_test.go
@@ -74,13 +74,13 @@ type frame struct {
 
 type wramSetValue struct {
 	// offset is relative to $7E0000, e.g. $F340 for bow
-	offset uint16
+	offset uint32
 	value  uint8
 }
 
 type wramTestValue struct {
 	// offset is relative to $7E0000, e.g. $F340 for bow
-	offset uint16
+	offset uint32
 	value  uint8
 }
